% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assesment.R
\name{anova_resample_test}
\alias{anova_resample_test}
\title{ANOVA test for models with in resamples}
\usage{
anova_resample_test(fitted_models, metric_name)
}
\arguments{
\item{fitted_models}{The Workflowsets}

\item{metric_name}{The accuracy metric to filter}
}
\value{
Pair-Wise comparison results (Tukeys)
}
\description{
ANOVA test for models with in resamples
}
