% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assesment.R
\name{bayesian_resample_test}
\alias{bayesian_resample_test}
\title{Bayesian Posterior Analysis for resamples}
\usage{
bayesian_resample_test(fitted_models, metric_name)
}
\arguments{
\item{fitted_models}{The Workflowsets}

\item{metric_name}{The accuracy metric to filter}
}
\value{
Posterior Distribution Values
}
\description{
Bayesian Posterior Analysis for resamples
}
