% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data_recipe}
\alias{data_recipe}
\title{The Data Recipe with Feature Engineering}
\usage{
data_recipe(DF, features)
}
\arguments{
\item{DF}{The Data to create Recipe}

\item{features}{The predictor variables to fit}
}
\value{
The recipe
}
\description{
The Data Recipe with Feature Engineering
}
