% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data_split}
\alias{data_split}
\title{Train/Test Split}
\usage{
data_split(DF, prop, resp)
}
\arguments{
\item{DF}{The Data Frame}

\item{prop}{The proportion of Training Data}

\item{resp}{The name of response variable}
}
\value{
rsplit
}
\description{
Train/Test Split
}
