% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_workflow.R
\name{model_workflow}
\alias{model_workflow}
\title{Create workflow sets}
\usage{
model_workflow(DF, model_name, features)
}
\arguments{
\item{DF}{The training Data Frame}

\item{model_name}{The model names}

\item{features}{The predictor variables to fit}
}
\value{
Workflow Set object
}
\description{
Create workflow sets
}
