% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visuals.R
\name{plot_error}
\alias{plot_error}
\title{Scatter Plot of Actual Vs. Predicted}
\usage{
plot_error(DF)
}
\arguments{
\item{DF}{The Data Frame}
}
\value{
Plot
}
\description{
Scatter Plot of Actual Vs. Predicted
}
