% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visuals.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot Histogram}
\usage{
plot_hist(DF, var, log_scale = FALSE)
}
\arguments{
\item{DF}{The Data Frame}

\item{var}{The Variable name}

\item{log_scale}{Boolean: Plot with log scale or not}
}
\value{
Plot
}
\description{
Plot Histogram
}
